[0.0.2] fixed subject
`subject` will be defined at global and therefore it does not make sense
to fail when trying to override it.

All the `subject` definition calls should be done before the first suite
is run and so there is no problem in overriding it. Also, the afterAll will
ensure that the original implementation is restored eventually.